# sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewDecision": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "APPROVE",
              "REJECT",
              "REQUEST_RECAPTURE"
            ],
            "title": "Action",
            "type": "string"
          },
          "expected_version": {
            "description": "The case version you reviewed (from GET /v1/review/{session}).",
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "note": {
            "description": "Why; stored encrypted, shown only to reviewers.",
            "maxLength": 1000,
            "minLength": 5,
            "title": "Note",
            "type": "string"
          },
          "reason_code": {
            "maxLength": 80,
            "minLength": 3,
            "pattern": "^[A-Z_]+$",
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "action",
          "reason_code",
          "note",
          "expected_version"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Server-side API key (kyc_\u2026).",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "Organization": {
        "description": "Organization UUID; required on every request.",
        "in": "header",
        "name": "X-Organization-ID",
        "type": "apiKey"
      },
      "ReviewerToken": {
        "description": "Expiring reviewer token (rvw_\u2026), separate from client credentials.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionClientToken": {
        "description": "Session client token (kst_\u2026) for capture and status of one session.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "KYC sessions, document extraction (Cambodian and international documents) MRZ and barcode validation, face quality, private 1:1 comparison, active liveness, ePassport chip verification, cross-checks, fraud signals and a deterministic risk engine, an authorized manual review dashboard, multi-tenant API keys with scopes, signed webhooks, security/privacy controls, and load-tested admission control (phases 1\u201318).",
    "title": "Universal Identity Platform",
    "version": "0.18.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/review/me": {
      "get": {
        "operationId": "me_v1_review_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Organization": [],
            "ReviewerToken": []
          }
        ],
        "summary": "Me",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/review/queue": {
      "get": {
        "description": "Cases waiting in MANUAL_REVIEW, longest-waiting first.",
        "operationId": "review_queue_v1_review_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Organization": [],
            "ReviewerToken": []
          }
        ],
        "summary": "Review Queue",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/review/{session_id}": {
      "get": {
        "description": "Everything this reviewer is authorized to see for one case (the view is audited).",
        "operationId": "review_case_v1_review__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Organization": [],
            "ReviewerToken": []
          }
        ],
        "summary": "Review Case",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/review/{session_id}/decision": {
      "post": {
        "description": "APPROVE, REJECT or REQUEST_RECAPTURE, with a reason code and a note (audited).",
        "operationId": "review_decision_v1_review__session_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Organization": [],
            "ReviewerToken": []
          }
        ],
        "summary": "Review Decision",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/review/{session_id}/images/{image_id}": {
      "get": {
        "description": "A retained document photo or selfie (VIEW_IMAGES; audited; never cached).",
        "operationId": "review_image_v1_review__session_id__images__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Organization": [],
            "ReviewerToken": []
          }
        ],
        "summary": "Review Image",
        "tags": [
          "review"
        ]
      }
    },
...
  },
  "security": [
    {
      "ApiKey": [],
      "Organization": []
    }
  ]
}